카테고리 아이템 페이징 개발#3

+++ src/main/java/org/example/tamaapi/dto/responseDto/MyPageable.java
@@ -1,6 +1,7 @@
 package org.example.tamaapi.dto.responseDto;
 
 import lombok.Getter;
+import org.springframework.data.domain.Pageable;
 
 @Getter
 public class MyPageable {
@@ -11,19 +12,22 @@ public class MyPageable {
     private int size;
 
     //계산된 페이지 수 (db row result / size)
-    private int pageCount;
-    /*
+    private long pageCount;
+
+
     //Page 객체는 0부터 시작 -> 1부터 시작하게 변경
     public MyPageable(Pageable pageable, long totalPages) {
         page = pageable.getPageNumber()+1;
         size = pageable.getPageSize();
         this.pageCount = totalPages;
     }
-     */
+
+
+    /* 직접 만든거
     public MyPageable( int page, int size, int rowCount) {
         this.page = page;
         this.size = size;
         this.pageCount = (int) Math.ceil((double) rowCount/size);
     }
-
+    */
 }
